+++ doc/whatsnew.rst
@@ -28,7 +28,7 @@ Various:
 - remove incorrectly spelled ``DonaichModel`` and ``donaich`` lineshape, deprecated in version 1.0.1 (PR #707)
 - remove occurrences of OrderedDict throughout the code; dict is order-preserving since Python 3.6 (PR #713)
 - update the contributing instructions (PR #718; @martin-majlis)
-
+- add check on input type of Model data and independent data. If the type is float32 a warning is generated (PR #723)
 
 .. _whatsnew_102_label: